Falling sand effect
I learned how to create a falling sand effect by clicking the mouse, just like the effect I wanted to achieve in quiz 9. When the mouse is clicked, sand particles of different colours fall, and when they touch particles of the same colour, they disappear. I want to add this interactive effect to our group code and integrate it with our group's artistic style.